@@ -14,14 +14,14 @@ android {
             storePassword '123456'
         }
     }
-    compileSdkVersion 33
+    compileSdkVersion 34
     defaultConfig {
       //  applicationId "com.wows.status"
         applicationId "com.wows.stats"
         minSdkVersion 19
-        targetSdkVersion 33
-        versionCode 29
-        versionName "1.29"
+        targetSdkVersion 34
+        versionCode 30
+        versionName "1.30"
         multiDexEnabled true
         signingConfig signingConfigs.config
     }
@@ -39,27 +39,28 @@ android {
     buildToolsVersion '33.0.0'
     productFlavors {
     }
+    namespace 'com.wows.status'
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     //noinspection GradleCompatible
-    implementation 'androidx.appcompat:appcompat:1.5.1'
-    implementation 'com.google.android.material:material:1.7.0'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.google.android.material:material:1.10.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation 'com.google.firebase:firebase-ads:21.4.0'
+    implementation 'com.google.firebase:firebase-ads:22.5.0'
     implementation 'com.google.firebase:firebase-core:21.1.1'
-    implementation 'com.google.firebase:firebase-inappmessaging-display:20.2.0'
+    implementation 'com.google.firebase:firebase-inappmessaging-display:20.4.0'
     implementation 'androidx.cardview:cardview:1.0.0'
-    implementation 'androidx.recyclerview:recyclerview:1.2.1'
-    implementation 'com.google.firebase:firebase-crashlytics:18.3.2'
-    implementation 'com.google.firebase:firebase-analytics:21.2.0'
+    implementation 'androidx.recyclerview:recyclerview:1.3.2'
+    implementation 'com.google.firebase:firebase-crashlytics:18.6.0'
+    implementation 'com.google.firebase:firebase-analytics:21.5.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'com.github.AnyChart:AnyChart-Android:1.1.2'
     implementation 'com.google.android.play:core:1.10.3'
-    implementation 'androidx.navigation:navigation-fragment:2.5.3'
-    implementation 'androidx.navigation:navigation-ui:2.5.3'
+    implementation 'androidx.navigation:navigation-fragment:2.7.5'
+    implementation 'androidx.navigation:navigation-ui:2.7.5'
 
 }
 
